#5: began work on output detection
