+++ crates/assemble-core/src/task/executable.rs
@@ -5,8 +5,8 @@ use crate::identifier::TaskId;
 use crate::project::buildable::{Buildable, BuiltByContainer, IntoBuildable};
 use crate::project::{ProjectError, ProjectResult, SharedProject};
 use crate::task::flags::{OptionDeclaration, OptionDeclarations, OptionsDecoder};
-use crate::task::previous_work::WorkHandler;
 use crate::task::up_to_date::{UpToDate, UpToDateContainer};
+use crate::task::work_handler::WorkHandler;
 use crate::task::{
     Action, BuildableTask, ExecutableTask, HasTaskId, TaskAction, TaskOrdering, TaskOrderingKind,
 };
