+++ crates/assemble-core/src/task/work_handler/output.rs
@@ -1,0 +1,8 @@
+use crate::file_collection::FileSet;
+use std::time::SystemTime;
+
+/// The output of a task
+pub struct Output {
+    timestamp: SystemTime,
+    files: FileSet,
+}
